+++ src/main/java/org/nhnacademy/lsj/Main.java
@@ -1,4 +1,4 @@
-package org.nhnacademy;
+package org.nhnacademy.lsj;
 
 import java.io.IOException;
 
@@ -13,9 +13,11 @@ public class Main {
 
         //Problem1.problem1();
         //Problem2.problem2();
-        //Problem3.problem3();
+        Problem3.problem3();
         //Problem4.problem4();
-        Problem5.problem5();
+        //Problem5.problem5();
+        //Problem6.problem6();
+        //Problem7.problem7();
 
     }
