+++ src/main/java/org/nhnacademy/lsj/Problem1.java
@@ -1,7 +1,5 @@
-package org.nhnacademy;
+package org.nhnacademy.lsj;
 
-import java.util.ArrayList;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
