+++ src/main/java/org/nhnacademy/lsj/Problem7.java
@@ -1,6 +1,5 @@
-package org.nhnacademy;
+package org.nhnacademy.lsj;
 
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Set;
 import org.slf4j.Logger;
@@ -21,7 +20,7 @@ public class Problem7 {
 
         int[] arr = new int[366];
         int count = 0;
-        while (true) {
+        while (true) { // if문을 while문 안에 넣어서 하기
             int birthDay = (int) (Math.random() * 365 + 1);
             arr[birthDay]++;
             count++;
@@ -34,7 +33,7 @@ public class Problem7 {
 
     public static void problem7_2() {
 
-        Set<Integer>set = new HashSet<>();
+        Set<Integer> set = new HashSet<>();
 
         for (int i = 0; i < 365; i++) {
             int birthDay = (int) (Math.random() * 365 + 1);
@@ -46,15 +45,15 @@ public class Problem7 {
 
     public static void problem7_3() {
 
-        Set<Integer>set = new HashSet<>();
-        long count=0;
+        Set<Integer> set = new HashSet<>();
+        long count = 0;
 
-        for (;;) {
+        while (true) { // if문 조건으로 쓰기
             int birthDay = (int) (Math.random() * 365 + 1);
             set.add(birthDay);
             count++;
-            if(set.size()==365){
-                logger.info("최소 한명 씩 찾기위한 시도횟수는 {}",count);
+            if (set.size() == 365) {
+                logger.info("최소 한명 씩 찾기위한 시도횟수는 {}", count);
                 return;
             }
         }
